refactor: OfferCard to separate file
to improve folder structure
